pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "automl-agent-opencode"
version = "0.1.0"
description = "AutoML-Agent OpenCode integration - multi-agent ML pipeline automation with Skills and MCP"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "AutoML Team", email = "team@example.com"}
]
keywords = ["automl", "machine-learning", "opencode", "agent", "ml-pipeline", "skills"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Machine Learning",
]
requires-python = ">=3.10"

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
opencode = [
    "mcp>=1.0.0",
]
core = [
    # Optional: Legacy core components
    # Install with: pip install -e ".[core]"
    # This will use the core/ directory if present
]
all = [
    "automl-agent-opencode[dev]",
    "automl-agent-opencode[opencode]",
]

[project.urls]
Homepage = "https://github.com/your-org/automl-agent-opencode"
Documentation = "https://automl-agent-opencode.readthedocs.io"
Repository = "https://github.com/your-org/automl-agent-opencode"
Issues = "https://github.com/your-org/automl-agent-opencode/issues"

[project.scripts]
automl-opencode = "adapter.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["adapter*", "mcp_servers*"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.json", "*.yaml"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short"]

[tool.coverage.run]
source = ["adapter"]
omit = ["*/tests/*", "*/mcp_servers/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
]
